Add documentation to all undocumented files
