+++ Image-Sorter-UI/app/Styles/ImageFolder.axaml.cs
@@ -1,12 +1,16 @@
 using System;
 using System.Net.Mime;
 using app.Model;
+using app.Views;
 using Avalonia;
 using Avalonia.Controls;
 using Avalonia.Controls.Primitives;
 
 namespace app.Styles;
 
+/// <summary>
+/// A style to display folders to be used on <see cref="AddImageDisplayView"/>
+/// </summary>
 public class ImageFolder : TemplatedControl
 {
     /// <summary>
