+++ Image-Sorter-UI/app/Styles/ObjectStyles.axaml
@@ -2,11 +2,9 @@
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
     <Design.PreviewWith>
         <Border Padding="20">
-            <!-- Add Controls for Previewer Here -->
         </Border>
     </Design.PreviewWith>
 
-    <!-- Add Styles Here -->
     <!-- button style -->
     <Style Selector="Button.baseButton">
         <Setter Property="HorizontalAlignment" Value="Stretch"/>
