+++ Image-Sorter-UI/app/ViewModels/AddImageDisplayViewModel.cs
@@ -23,43 +23,7 @@ public class AddImageDisplayViewModel: ViewModelBase, IAddImageDisplayViewModel
     /// </summary>
     private bool _folderPrompt;
 
-    // TODO This is a little bit of a hack maybe refactor at a later date
-    /// <summary>
-    /// Checks if file is already in <see cref="MainWindowViewModel.FolderList"/>,
-    /// also checks that folder is not a child of any folders already
-    /// present in <see cref="MainWindowViewModel.FolderList"/>, removes child folders from
-    /// <see cref="MainWindowViewModel.FolderList"/> if folder being added is a parent
-    /// </summary>
-    /// <param name="folder">The folder being searched for</param>
-    /// <returns>True if folder is not already present in <see cref="MainWindowViewModel.FolderList"/> and is not a child</returns>
-    private bool IsNotPresent(SelectFolders folder)
-    {
-        var children = new List<SelectFolders>();
-        var folderPath = EndsWithSeparator(folder.Path);
-        foreach (var presentFolder in MainModel!.FolderList)
-        {
-            var presentPath = EndsWithSeparator(presentFolder.Path);
-            if (presentPath.StartsWith(folderPath, StringComparison.OrdinalIgnoreCase)) children.Add(presentFolder);
-            else if (folderPath.StartsWith(presentPath, StringComparison.OrdinalIgnoreCase)) return false;
-        }
-
-        foreach (var child in children)
-        {
-            MainModel!.RemoveFolders(child);
-        }
-
-        return true;
-    }
     
-    /// <summary>
-    /// Removes separators from the end of file paths
-    /// </summary>
-    /// <param name="absolutePath">A <see cref="string"/> of the absolute path of a <see cref="Uri"/></param>
-    /// <returns></returns>
-    private static string EndsWithSeparator(string absolutePath)
-    {
-        return absolutePath.TrimEnd('/','\\') + "/";
-    }
 
     /// <summary>
     /// Checks if a featureGroup has any items selected
@@ -159,15 +123,10 @@ public class AddImageDisplayViewModel: ViewModelBase, IAddImageDisplayViewModel
         }
     }
 
-
     /// <inheritdoc/>
     public void AddFolders(List<SelectFolders> folders)
     {
-        foreach (var folder in folders.Where(IsNotPresent))
-        {
-            MainModel?.FolderList.Add(folder);
-        }
-
+        MainModel?.AddFolders(folders);
         this.RaisePropertyChanged(nameof(FoldersEmpty));
     }
