+++ Image-Sorter-UI/app/ViewModels/Interfaces/IMainWindowViewModel.cs
@@ -1,9 +1,13 @@
+using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using app.Model;
 using app.Views;
 
 namespace app.ViewModels.Interfaces;
 
+/// <summary>
+/// The ViewModel of the <see cref="MainWindow"/>
+/// </summary>
 public interface IMainWindowViewModel
 {
    /// <summary>
@@ -14,7 +18,7 @@ public interface IMainWindowViewModel
     /// <summary>
     /// Holds the current display, when changed page shown will change
     /// </summary>
-    public ViewModelBase CurrentPage { get; set; }
+    public ViewModelBase CurrentPage { get; }
     
     /// <summary>
     /// Is true if the <see cref="CurrentPage"/> is set to that of <see cref="AddImageDisplayView"/>
@@ -30,7 +34,7 @@ public interface IMainWindowViewModel
     /// Adds folders to <see cref="FolderList"/>
     /// </summary>
     /// <param name="folders">the folders being passed by the start processing button</param>
-    public void AddFolders(ObservableCollection<SelectFolders> folders);
+    public void AddFolders(IEnumerable<SelectFolders> folders);
 
     /// <summary>
     /// Remove folders from the <see cref="FolderList"/>
