+++ Image-Sorter-UI/app/ViewModels/ViewModelProvider.cs
@@ -2,6 +2,7 @@ using app.ViewModels.Interfaces;
 
 namespace app.ViewModels;
 
+/// <inheritdoc cref="IViewModelProvider"/>
 public sealed class ViewModelProvider : IViewModelProvider
 {
     /// <summary>
