+++ Image-Sorter-UI/app/Views/AddImageDisplayView.axaml.cs
@@ -1,6 +1,4 @@
 using System;
-using System.Collections.Generic;
-using System.IO;
 using System.Linq;
 using app.Model;
 using app.ViewModels.Interfaces;
