+++ Image-Sorter-UI/app/Views/FolderStuctureDisplayView.axaml
@@ -3,6 +3,7 @@
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              xmlns:vm="clr-namespace:app.ViewModels.Interfaces"
+             xmlns:m="clr-namespace:app.Model"
              xmlns:styles="clr-namespace:app.Styles"
              mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
              x:DataType="vm:IFolderStructureDisplayViewModel"
@@ -50,10 +51,10 @@
                                   RectangleWidth="230" />
             </Grid>
             <ScrollViewer Grid.Row="1">
-                <ItemsControl ItemsSource="{Binding Directories}">
+                <ItemsControl  ItemsSource="{Binding Directories}">
                     <ItemsControl.ItemTemplate>
                         <DataTemplate>
-                            <styles:FolderOrdering Text="{Binding Key}" Height="40"/>
+                            <styles:FolderOrdering Text="{Binding }" Height="40"/>
                         </DataTemplate>
                     </ItemsControl.ItemTemplate>
                 </ItemsControl>
