+++ Image-Sorter-UI/app/Views/MainWindow.axaml.cs
@@ -1,9 +1,11 @@
 using app.ViewModels.Interfaces;
-using Avalonia.Controls;
 using Avalonia.ReactiveUI;
 
 namespace app.Views;
 
+/// <summary>
+/// A view which hosts all subviews within it
+/// </summary>
 public partial class MainWindow : ReactiveWindow<IMainWindowViewModel>
 {
     public MainWindow()
